refactor(wireguard): update import paths for peer IP reconciliation utility
