+++ app/operation/group.py
@@ -26,7 +26,7 @@ from app.models.group import (
 from app.node.sync import sync_users
 from app.operation import BaseOperation, OperatorType
 from app.utils.logger import get_logger
-from app.utils.wireguard import reconcile_wireguard_peer_ips_for_users
+from app.utils.wireguard_reconcile import reconcile_wireguard_peer_ips_for_users
 
 logger = get_logger("group-operation")
